First pass on biomes!
Rewrote the set_cell_with_noise to include humidity and temperature values for procedural generation.

Cleaned up the code for set_cell_with_noise to be easier to understand what it is doing and to modify later.

Map persistence (loading/saving to a file) is the next step needed to prepare for further development of the map generation.